Mn dataset

+++ camera_drive.py
@@ -56,7 +56,7 @@ class SipeedCamera(object):
         self.connection.write(str.encode(f'tools.measure_latency("{model_file}")'))
         self.connection.write(b'\r')
         time.sleep(1)
-        self.wait_to_ready(timeout=12000)  # 2min
+        self.wait_to_ready(timeout=3000)  # 30s
 
         self.connection.write(b'tools.query_latency()')
         self.connection.write(b'\r')
@@ -71,25 +71,32 @@ class SipeedCamera(object):
         df = pd.read_csv(f"{path}/table.csv")
 
         for h5 in h5_list:
+            kmodel_index = df[df["model"] == h5.split(".")[0]].index
             file_path = f"{path}/{h5}".split(".")[0]
-            if not os.path.isfile(file_path + ".kmodel"):
+            if not os.path.isfile(file_path + ".kmodel") and pd.isna(df.loc[kmodel_index[0]]["kmodel_memory [KB]"]):
+                print(f"Converting {file_path} ..")
                 converter = tf.compat.v1.lite.TFLiteConverter.from_keras_model_file(file_path + ".h5")
                 tflite_model = converter.convert()
                 open(file_path + ".tflite", "wb").write(tflite_model)
                 # os.system(f"./nncase/ncc compile {file_path}.tflite {file_path}.kmodel -i tflite -o kmodel "
                 #           f"--dataset nncase/calibration_dataset")
-                terminal_out = subprocess.run(["./nncase/ncc", "compile", f"{file_path}.tflite", f"{file_path}.kmodel",
-                                               "-i", "tflite", "-o", "kmodel", "--dataset",
-                                               "nncase/calibration_dataset"],
-                                              stdout=subprocess.PIPE)
+                try:
+                    terminal_out = subprocess.run(["./nncase/ncc", "compile", f"{file_path}.tflite",
+                                                   f"{file_path}.kmodel", "-i", "tflite", "-o", "kmodel", "--dataset",
+                                                   "nncase/calibration_dataset"], stdout=subprocess.PIPE, timeout=180)
+                except subprocess.TimeoutExpired as _:
+                    print("kmodel conversion timeout.")
+                    terminal_out = "Working memory usage: 0"  # 0 means failed
                 pattern = re.compile(r"Working memory usage: (\d+)")  # number after a text
                 match = re.findall(pattern, str(terminal_out))
                 if len(match):
                     kmodel_memory = round(int(match[0])/1000, 2)
-                    kmodel_index = df[df["model"] == h5.split(".")[0]].index
                     df.at[kmodel_index, "kmodel_memory [KB]"] = kmodel_memory
+                    df.to_csv(f"{path}/table.csv", index=False)
+                    print(f"Converting {file_path} done.")
                 else:
+                    df.at[kmodel_index, "kmodel_memory [KB]"] = 0
+                    df.to_csv(f"{path}/table.csv", index=False)
                     print(f"Converting {file_path} failed.")
-                print(f"Converting {file_path} done.")
 
         df.to_csv(f"{path}/table.csv", index=False)
